requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.13.1
fsspec==2024.2.0
gradio==5.0.1
gradio_client==1.4.0
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rich==13.9.2
ruff==0.6.9
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
swarmauri==0.4.1
swarmauri-core==0.5.0
sympy==1.12
tomlkit==0.12.0
torch==2.4.1+cpu
torchaudio==2.4.1+cpu
torchvision==0.19.1+cpu
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
websockets==12.0
